+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+- Added outlines / bookmarks for `text` node
+	- `outline` - set to `true` for add to bookmarks
+	- `outlineText` (optional) - set custom bookmark text, otherwise text from node
+	- `outlineExpanded` (optional) - set to `true` for expanded/opened bookmark
+	- `outlineParentId` (optional) - parent bookmark `id`
+
 ## 0.3.4 - 2026-02-13
 
 - Added vertical alignment for table cells via `verticalAlignment` property, values: `top` (default), `middle`, `bottom`